// frontend/src/api/model/courseResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface CourseResponse {
  id?: string;
  name?: string;
  description?: string;
  inviteCode?: string;
  ownerId?: string;
  ownerName?: string;
  createdAt?: string;
  updatedAt?: string;
}